[fix] corrigindo regra para transações fixas e exibição de elementos

+++ src/app/pages/entradas/entradas.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnChanges, OnInit, SimpleChanges, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatPaginator, PageEvent } from '@angular/material/paginator';
 import { ReceitasModalComponent } from './receitas-modal/receitas-modal.component';
 import { MatDialog } from '@angular/material/dialog';
@@ -8,7 +8,6 @@ import { Transacao } from 'src/app/shared/models/transacao.model';
 import { IDate } from 'src/app/shared/models/date.interface';
 import { takeUntil } from 'rxjs';
 import { OnDestroyService } from 'src/app/shared/service/onDestroy/on-destroy.service';
-import { v4 as uuidv4 } from 'uuid';
 import { IUser } from 'src/app/shared/models/user.model';
 import { AuthService } from 'src/app/core/services/auth/auth.service';
 
@@ -66,7 +65,14 @@ export class EntradasComponent extends OnDestroyService implements OnInit {
   }
 
   removeTransaction(transacao: Transacao) {
-    this._transactionService.deleteTransaction(transacao);
+
+    if(!transacao.transacao_repetida && transacao.parentId != null) {
+      let deletedTransaction: Transacao = {...transacao, ativo: false, dataExclusao: new Date().toISOString() };
+      this._transactionService.updateTransaction(deletedTransaction);
+    } else {
+      this._transactionService.deleteTransaction(transacao);
+    }
+
     this.getFinancialIncomeByDate(this.currentYear, this.currentMonth, String(this.user?.id), "receita", null);
 
     if(this.todasTransacoes.length == 0) {
@@ -96,7 +102,7 @@ export class EntradasComponent extends OnDestroyService implements OnInit {
   changeYear(newYear: number) {
     this.currentYear = newYear;
 
-    this.gerarTransacoesFixas(this.currentMonth, this.currentYear);
+    this._transactionService.generateFixedTransactions("receita", String(this.user?.id), this.currentYear, this.currentMonth);
     this.getFinancialIncomeByDate(this.currentYear, this.currentMonth, String(this.user?.id), "receita", null);
 
     if (this.paginator) {
@@ -109,7 +115,7 @@ export class EntradasComponent extends OnDestroyService implements OnInit {
     this.listaTransacoes = [];
     this.currentMonth = newMonth.index;
     
-    this.gerarTransacoesFixas(this.currentMonth, this.currentYear);
+    this._transactionService.generateFixedTransactions("receita", String(this.user?.id), this.currentYear, this.currentMonth);
     this.getFinancialIncomeByDate(this.currentYear, this.currentMonth, String(this.user?.id), "receita", null);
     
     if (this.paginator) {
@@ -125,7 +131,7 @@ export class EntradasComponent extends OnDestroyService implements OnInit {
       this.currentYear = this.currentYear + 1;
     }
 
-    this.gerarTransacoesFixas(this.currentMonth, this.currentYear);
+    this._transactionService.generateFixedTransactions("receita", String(this.user?.id), this.currentYear, this.currentMonth);
     this.getFinancialIncomeByDate(this.currentYear, this.currentMonth, String(this.user?.id), "receita", null);
 
     if (this.paginator) {
@@ -142,6 +148,7 @@ export class EntradasComponent extends OnDestroyService implements OnInit {
       this.currentYear = this.currentYear - 1;
     }
 
+    this._transactionService.generateFixedTransactions("receita", String(this.user?.id), this.currentYear, this.currentMonth);
     this.getFinancialIncomeByDate(this.currentYear, this.currentMonth, String(this.user?.id), "receita", null);
 
     if (this.paginator) {
@@ -160,39 +167,6 @@ export class EntradasComponent extends OnDestroyService implements OnInit {
     }
 
     this.getFinancialIncomeByDate(this.currentYear, this.currentMonth, String(this.user?.id), "receita", this.lastVisibleItem);
-
-  }
-
-  private gerarTransacoesFixas(mes: number, ano: number) {
-    const transacoesFixas = this.todasTransacoes.filter((transacao: Transacao) => transacao.transacao_fixa );
-
-    if(transacoesFixas.length == 0) {
-      return;
-    }
-
-    transacoesFixas.forEach((transacaoFixa: Transacao) => {
-      let newDate: any = new Date(transacaoFixa.data);
-      const currentDate = new Date(this.currentYear, this.currentMonth, new Date().getDate());
-  
-      if(currentDate >= newDate) {
-        newDate = newDate.setFullYear(this.currentYear, this.currentMonth, newDate.getDate())
-
-        this._transactionService.getTransactionsById(transacaoFixa.id, new Date(newDate).getMonth(), new Date(newDate).getFullYear()).pipe(takeUntil(this.destroy$)).subscribe(transacao => {
-          if(transacao.length == 0) {
-            const transacao: Transacao = {
-              ...transacaoFixa,
-              data: new Date(newDate).toISOString(),
-              ano: new Date(newDate).getFullYear(),
-              mes: new Date(newDate).getMonth(),
-              uniqueId: uuidv4()
-            }
-            
-            this._transactionService.createTransaction(transacao);
-          }
-          
-        });
-      }
-    });
 
   }
 }
